+++ src/main.js
@@ -1,7 +1,8 @@
 import {EVENT_COUNT, RENDER_POSITION} from "./const";
 import {TripInfoView, NavigationControllerView, EventFiltrationView, TotalPriceView, SortingView, EventEditorView, DaysView, DayView, EventView, NoEventView} from "./view/index";
 import {generateEvent, DESTINATIONS} from "./mock/event";
-import {splitEventsByDays, renderElement} from "./utils";
+import {splitEventsByDays} from "./utils/event";
+import {renderElement} from "./utils/render";
 
 const events = new Array(EVENT_COUNT).fill().map(generateEvent).sort((a, b) => a.dateStart - b.dateStart);
