+++ src/utils/render.js
@@ -1,0 +1,22 @@
+import {RENDER_POSITION} from "../const";
+
+export const renderElement = (container, element, place) => {
+  switch (place) {
+    case RENDER_POSITION.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RENDER_POSITION.BEFOREEND:
+      container.append(element);
+      break;
+    case RENDER_POSITION.AFTEREND:
+      container.insertAdjacentElement(place, element);
+      break;
+  }
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
