+++ src/view/event-editor.js
@@ -1,6 +1,6 @@
 import {EVENT_TYPE_TRANSFER, EVENT_TYPE_ACTIVITY} from "../const";
 import AbstractView from "./abstract";
-import {getDateAtDefaultFormat, getTimeAtDefaultFormat} from "../date-formatters";
+import {getDateAtDefaultFormat, getTimeAtDefaultFormat} from "../utils/date-formatters";
 import {getOffers} from "../mock/event";
 
 const BLANK_EVENT = {
